Native external hbase table access (select, IUD) changes.
-- IUD on external hbase tables is now enabled by default
-- predicates on native hbase tables can now be pushed down to
   hbase region server
-- traf varchar col maxlength is now 200K by default,
   can be changed by cqd max_character_col_size
-- executor handles column values length greater than 32K during
   move to/from JNI
-- error is correctly returned if data retrieved from hbase exceeds expected
   max row length
-- hbase column_create function now takes an expression/param as its
   column name operand

+++ sql/executor/ExHbaseAccess.h
@@ -317,7 +317,7 @@ protected:
   Lng32 copyColToDirectBuffer( BYTE *rowCurPtr, 
                 char *colName, short colNameLen,
                 NABoolean prependNullVal, char nullVal, 
-                char *colVal, short colValLen);
+                char *colVal, Int32 colValLen);
   short copyRowIDToDirectBuffer(HbaseStr &rowID);
 
   short numColsInDirectBuffer()
